Commit do Breno Vinicius

+++ deletar.php
@@ -6,15 +6,20 @@
     $senha = md5($senha);
     
     $conecta = mysqli_connect('sql10.freesqldatabase.com','sql10348436','3CLi39vjV7','sql10348436', '3306');
+    $tenta_achar = "SELECT * FROM usuarios WHERE email='$email' AND senha='$senha'" ;
+    $resultado = $conecta->query($tenta_achar);
+    $user = mysqli_fetch_assoc($resultado);
+    if ($resultado->num_rows > 0) {
+        $sql = "DELETE FROM usuarios  WHERE email ='$email'"; 
     
-   $sql = "DELETE FROM usuarios  WHERE email ='$email'";     
-     if ($conecta->query($sql) === TRUE) 
-         {         
-            echo"<script language='javascript' type='text/javascript'>
-          alert('Conta excluida com sucesso!');window.location.
-          href='index.php'</script>";     
+        if ($conecta->query($sql) === TRUE) 
+            {         
+                echo"<script language='javascript' type='text/javascript'>
+                alert('Conta excluida com sucesso!');window.location.
+                href='index.php'</script>";     
          
-         } 
+            } 
+    }
         else 
         {         
             echo"<script language='javascript' type='text/javascript'>
